Still a work in progress, PopupDialog.java was created to generate informational or input style popup dialogs with the basic requirements of a context, title and message (and single input field if needed). Globals.java was created to manage manage the List<> identification of note data similar to enums where a name represents a number (e.g., TYPE = 0, SOURCE = 2, etc.). This may later just be merged to the GlobalVariables.java or visa versa. FileChooser.java was created to search and add files to be attached to a note.
Many edits, refactoring and fixes were made such as preventing the edit/update of an existing note source, with the expectation the source to be accurate and the only data to be updated from the source would be pages and timestamps only. Updates to the AddNote.java and corrections to the ViewNote.java and EditNote.java, which were accidentally broken in managing the date field. Many other various updates and changes made.