DateTimestampManager.java was introduced into the app to help check and validate acceptable date and timestamp formats.
New DBQueryTools includes new methods yet to be defined

Many edits, refactoring and fixes were made such as preventing the edit/update of an existing note source, with the expectation the source to be accurate and the only data to be updated from the source would be pages and timestamps only. Updates to the AddNote.java and corrections to the ViewNote.java and EditNote.java, which were accidentally broken in managing the date field. Many other various updates and changes made.