Numerous development changes and additions just to manage Sources and Authors during the Adding of new notes. In total, there were 31 modifications and 7 new additions (e.g., classes or resources).
Many classes were modified or created. All developments since the last code commit were to manage new and existing sources and authors based upon the user choices. It was assumed that some titles could be duplicated since multiple authors could have books with the same title, for example, "The Case for Christianity" is shared between at least two popular authors if not more self-published authors less known.

Nearly all Android Room database DAO's INSERT methods were modified to return a long number which were previously void methods. It was learned that rather than have an annotated method to return the last table id created, the INSERT method(s) were defined in room to return either a long or void - determined by the developer.

Nearly all classes were modified to in some capacity to better manage the source and author(s) pertaining to them. New classes such as PopupAdapter, AuthorPopup and DataHolder were to help in prompting the user to select the correct author(s) on like titles, or whether that title was published by a new, unknown author not currently in the database.

As a result of new classes, new activities were created and resource values added.

Issues remaining on the subject of sources and authors:

At the time of the commit, an AddNote.java issue between the sourceTitle and tableLayoutAuthors listeners to prevent duplicate triggering exists. Both the listeners trigger populateSourceDetails method, which manages and aligns the correct authors and source on all titles, or whether the user needs to make a choice between authors, or the source is new.

There also exists the issue of writing data to the database in the case of new titles with new and or existing authors.