Updates to EditNote.java, activity_edit_note.xml, AddNote.java, arrays.xml, and strings.xml
Due to issues with database versioning and updates, the project at this point was restored from an 08/24/2021 copy from GitHub. After confirmation that the app could launch and use the database successfully, updates not pertaining to the database were made to the files above made between 08/24/2021 and 09/02/2021.

EditNote / activity_edit_note.xml were revised to prevent updating the following fields: Topic, Source, Author, Date, Volume, Edition, and Issue. Reasoning: the user should not need to modify these if sources are entered correctly. If sources are incorrectly identified, the note should be deleted, which should remove any source not currently linked to another note and allow the source to be entered correctly.

AddNote, strings.xml and arrays.xml were modified to manage the passing of a new note to the DBQueryTools.

Current Issues:
There are still issue with updating Notes in the database with "0" values and so modifications will need to be made to the database and Android Room structure around the Notes Entity to allow these updates to be made to the Notes join table for non-essential table references. The previous attempt to make these modifications was the reason requiring a restoration.