The NotesDao.java, TermsDao.java, and TopicsDao.java were updated with new methods to return primary IDs by their string values and to return counts based upon one or both counts by string and by ID.
The DateTimestampManager.java and AddNote.java method to captureNoteDetails was modified to add the data parameters in the correct order. The AddNote.java included an additional requiredFields() check to validate new sources have authors listed in the Authors table.

The DBQueryTools.java had 4 updated methods added to efficiently manage the updates to Questions, Quotes, Terms and Topics during the note update procedure.

The MainActivity.java topic, question and custom fields were updated to null and changed to obj.setText() where appropriate.

The activity_add_note.xml was modified as a text input to prevent the sources field defaulting to a Return on the keyboard vs Done. This helps prevent the user from accidentally adding a return in a source title and allows the title to wrap or stay on a single line of text.

The strings.xml was updated with a new string hint for the custom search field.