Issues started occurring with the ability to delete the table rows for both the File and Author tables. The updated deleting function was impacting this ability. Uploaded a previous copy of the BuildTableLayout.java, which seemed to correct the issue.
This update includes a new "Add File" button to take over the function of the add button of the Header. There was an issue due to lack of understanding and the FileChooser to work properly. A separate button was created to manage and launch the FileChooser prior to creating the new row, then populate the file path and name.