The Apologetic.db was an updated pre-populated db created from the ResearchDB desktop version to make sure it would be interchangeable/readable with the phone version.
BuildTableLayout.java methods were updated to "fix" columns within TableLayouts. There was a missing LayoutParams declaration for TableRows created which contributed to column rows with lengthy values causing forcing adjacent columns "disappear" so-to-speak and as a result were not accessible. When TableRows for the files TableLayout, tags were created to help identify the correct rows to delete. As a result, the BuildTableLayout.deleteTableRows method was updated to iterate through the child rows of the TableLayout to find a TextView string value matching the row tag, if so, the TableRow view (child of the table) was then deleted (removed). There may need to be adjustments in this functionality when authors are added and removed at a later time. All of these updates impacted the AddNote.java and EditNote.java which in turn were updated.

DBQueryTools.java methods were updated to identify the differences between previous and new List<Files> to determine which files between the two were either to be ignored, removed, or added. Ignoring would mean the file was to remain, whereas the others were simply added or removed. This will need to be revisited and probably refactored to be more efficient and possibly less cumbersome in resource dependency.