Attempting a second commit to resolve a 'red filename issue'.
Added the MasterDatabase and related files (e.g., classes, daos, entities, etc.) to begin the implementation of multiple database creation and management. Uncertain at this point whether the new database management will incorporate a pre-populated database sample or not. The point of this commitment is to update the project in git.

The directory and file structure were updated to more clearly manage the distinction between the MasterDatabase and ResearchDatabase daos and entities.