Fixed an issue in which the switching of databases wasn't changing because the first initialization of the Room Database instance variable would remain; and as a result, when switching to a different database, this variable was not null and the null check would be skipped wrapping the database instantiation. There were two possible solutions, reinitialize the variable everytime or null the variable and reinitialize every time.
Implemented the ability to import databases from the desktop version or other "ResearchDB Apps."

Implemented a checkTopic function in the DBQueryTools.java to check whether a new topic was created to determine whether or not the MainActivity.java should refresh on returning. Additional notes on a topic would appear upon returning but a new topic would not.