custom_database_list.xml was modified to remove the red "x" and substitute an RDB logo. There is no purpose except the replacement with the icon indicates 1) it is a database and 2) it looks good.
Modified the ListView for the DatabaseList to open an the selected database upon clicking verse an extra step of clicking on the button. The button was purposeful in confirming the selection made but most would probably find it a waste.

Included keyboard closing functionality that assured the keyboard closed upon a dropdown seletion or upon the use of other logical views vs remaining open and requiring the user the extra effort to close separately.

Modified the layout of the MainActivityMaster xml expanding the allowed database name length from 17(20) to 25(28) characters (allowing for 3 characters for .db). Adjusted the layout of the AddDatabase button and removed the useSelectedDatabase button with the modification the ListView onClick for each item.