Implemented DBQueryTools.deleteNote method to delete notes, which must be marked for delete to be deleted. Additionally, if the note is marked for delete, it won't show up in the regular list of notes when queried. Under the MainActivity menu, the option to review notes is the only method to reselect the note for deletion or to delete permanently. A note cannot be deleted on the search page, it must be viewed to delete it or un-mark it for delete. The cumbersome way is intentionally to make it difficult to delete a note accidentally the user does not want to delete. Much like deleting a database cannot be deleted from the database page (i.e., MainActivityMaster).
Custom classes for OnFocusChangeListener and OnKeyListener were created to make the classes generic and allow single line calls in the Activities.

Menu updates were made to the main_menu.xml and add_menu.xml to eliminate unnecessary clicking for the user. Categorically it made sense, but with few menu options, categories for some options didn't.