AddNote.java was updated to zero an existing SourceID to allow a user to add duplicating source titles since it is possible to have duplicate titles with different authors.
DBQueryTools.captureDBSources included an "if" statement to remove any duplicate titles in the case of a title having more than one separate author resulting in two distinct sources with the same title.

Prior code was implemented to prompt the user to choose a specific author when duplicate titles exist, or to click "New Source" with in the prompt when a third may exist and not yet added.