Added note importing validation to prevent crashing but informing the user of invalid xml files.
Added database importing validation to prevent un-compatible databases from being imported.

The Delete Database portion of the main_menu.xml was moved to a submenu of Delete Database to make it difficult to easily delete a database, require one extra step. There still is a message prompt to validate the intentions too.