Fixed a bug involving the Authors table when a row is added manually by the user. The Author EditText field should populate with the default message telling the user to add authors in the Authors table. However, this was not happening because the added author row was not given focus. As a result, various related AddNote methods and BuildTableLayout.setupAuthorsAddRowButton were modified to receive focus immediately after the row creation. This resolved the issue.
Fixed a bug within the EditNote.java in which the page and timestamp would not update when changed. This was occurring because the new field values were not being passed to the updateNoteDetails list. Some source fields are intentionally un-editable (e.g., Author, Date, Volume, Edition, Issue) which should never change in value and the page and timestamp fields were accidentally included. These values can change because different pages, paragraphs or timestamps can be referenced from the same source. As a result, these fields should not be fixed in value.