An issue with screen orientation was causing data loss. Update to the AndroidManifest.xml adding configChanges and alwaysRetainTaskState resolved the issue.
Updated Topic hint and corrected SourcesDao to list titles alphabetically

Updated the management and passing of a file Uri. Due to Android 11 (API 30), I learned that to properly circumvent the use of the download folder, the getContentResolver.openInputStream() can (apparently) only receive the Uri directly and not be converted from the Uri.fromFile().